Move `CoreSearchRequest` and `QueryEnum` from `collection` into `shard` crate
