+++ lib/collection/src/operations/mod.rs
@@ -6,7 +6,6 @@ pub mod conversions_rest;
 pub mod operation_effect;
 pub mod payload_ops;
 pub mod point_ops;
-pub mod query_enum;
 pub mod shard_selector_internal;
 pub mod shared_storage_config;
 pub mod snapshot_ops;
@@ -17,6 +16,10 @@ pub mod validation;
 pub mod vector_ops;
 pub mod vector_params_builder;
 pub mod verification;
+
+pub mod query_enum {
+    pub use shard::search::QueryEnum;
+}
 
 use std::collections::HashMap;
