+++ lib/collection/src/operations/verification/operation_rate_cost.rs
@@ -1,6 +1,6 @@
 use segment::types::Filter;
 
-use crate::operations::types::{CoreSearchRequest, QueryScrollRequestInternal};
+use crate::operations::types::QueryScrollRequestInternal;
 
 pub fn filter_rate_cost(filter: &Filter) -> usize {
     filter.total_conditions_count()
@@ -8,16 +8,6 @@ pub fn filter_rate_cost(filter: &Filter) -> usize {
 
 /// Base cost for a read operation
 pub const BASE_COST: usize = 1;
-
-impl CoreSearchRequest {
-    pub fn search_rate_cost(&self) -> usize {
-        let mut cost = self.query.search_cost();
-        if let Some(filter) = &self.filter {
-            cost += filter_rate_cost(filter);
-        }
-        cost
-    }
-}
 
 impl QueryScrollRequestInternal {
     pub fn scroll_rate_cost(&self) -> usize {
