+++ lib/shard/src/lib.rs
@@ -2,6 +2,7 @@ pub mod locked_segment;
 pub mod operations;
 pub mod payload_index_schema;
 pub mod proxy_segment;
+pub mod search;
 pub mod segment_holder;
 pub mod update;
 pub mod wal;
